+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 ## 2.1.0 — Unreleased
 
 * Update the Pusher JS client to version 7.0.0. (Tristan Dunn)
+* Update development and test dependencies. (Tristan Dunn)
 
 ## 2.0.0 — March 31, 2020